added absolute ticks for F9R integration

+++ vesc_driver/include/vesc_driver/vesc_packet.h
@@ -191,7 +191,7 @@ public:
   double getConsumedPower() const;
   double getInputPower() const;
   uint32_t getPosition() const;
-  double getDisplacement() const;
+  uint32_t getDisplacement() const;
   int getFaultCode() const;
 
 private:
